# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  session.ServerSessionRequest:
    properties:
      duration:
        type: string
      server_group:
        type: string
    type: object
  session.ServerSessionResponse:
    properties:
      duration:
        type: string
      expiry:
        type: string
      server_group:
        type: string
    type: object
...
info:
  contact: {}
  description: 'All endpoints return responses wrapped in a standard envelope: {"success":
    bool, "data": (schema shown per endpoint), "error": string}. All protected endpoints
    can return 401.'
  title: ez2boot API
  version: "1.0"
paths:
  /admin/session:
    put:
      consumes:
      - application/json
      description: '[Admin] modify a server session.'
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/session.ServerSessionRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/session.ServerSessionResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Update server session admin
      tags:
      - session
...
  /session:
    post:
      consumes:
      - application/json
      description: Create a new server session
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/session.ServerSessionRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/session.ServerSessionResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: New server session
      tags:
      - session
    put:
      consumes:
      - application/json
      description: Modify own server session.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/session.ServerSessionRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/session.ServerSessionResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Update server session
      tags:
      - session
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"